+++ _color set generator/_output/qmc_array.py
@@ -1,0 +1,467 @@
+# QMC SELECT COLORS SET
+
+import bpy
+
+from .globals import *
+from .color_functions import *
+
+# QMC SELECT OPERATORS
+
+class INTERNATIONAL_ORANGE_ENGINEERING(bpy.types.Operator):
+    """International Orange Engineering"""
+    bl_label = "International Orange Engineering"
+    bl_idname = 'color.international_orange_engineering'
+    def execute(self, context):
+        set_base_color(0xBA160C, self.bl_label)
+        return {'FINISHED'}
+
+class SCARLET(bpy.types.Operator):
+    """Scarlet"""
+    bl_label = "Scarlet"
+    bl_idname = 'color.scarlet'
+    def execute(self, context):
+        set_base_color(0xFF2000, self.bl_label)
+        return {'FINISHED'}
+
+class VERMILION(bpy.types.Operator):
+    """Vermilion"""
+    bl_label = "Vermilion"
+    bl_idname = 'color.vermilion'
+    def execute(self, context):
+        set_base_color(0xFF4000, self.bl_label)
+        return {'FINISHED'}
+
+class INTERNATIONAL_ORANGE_AEROSPACE(bpy.types.Operator):
+    """International Orange Aerospace"""
+    bl_label = "International Orange Aerospace"
+    bl_idname = 'color.international_orange_aerospace'
+    def execute(self, context):
+        set_base_color(0xFF4F00, self.bl_label)
+        return {'FINISHED'}
+
+class PERSIMMON(bpy.types.Operator):
+    """Persimmon"""
+    bl_label = "Persimmon"
+    bl_idname = 'color.persimmon'
+    def execute(self, context):
+        set_base_color(0xEC5800, self.bl_label)
+        return {'FINISHED'}
+
+class PERSIAN_ORANGE(bpy.types.Operator):
+    """Persian Orange"""
+    bl_label = "Persian Orange"
+    bl_idname = 'color.persian_orange'
+    def execute(self, context):
+        set_base_color(0xD99058, self.bl_label)
+        return {'FINISHED'}
+
+class BUFF(bpy.types.Operator):
+    """Buff"""
+    bl_label = "Buff"
+    bl_idname = 'color.buff'
+    def execute(self, context):
+        set_base_color(0xE0AB76, self.bl_label)
+        return {'FINISHED'}
+
+class COSMIC_LATTE(bpy.types.Operator):
+    """Cosmic Latte"""
+    bl_label = "Cosmic Latte"
+    bl_idname = 'color.cosmic_latte'
+    def execute(self, context):
+        set_base_color(0xFFF8E7, self.bl_label)
+        return {'FINISHED'}
+
+class AMBER(bpy.types.Operator):
+    """Amber"""
+    bl_label = "Amber"
+    bl_idname = 'color.amber'
+    def execute(self, context):
+        set_base_color(0xFFBF00, self.bl_label)
+        return {'FINISHED'}
+
+class SAFFRON(bpy.types.Operator):
+    """Saffron"""
+    bl_label = "Saffron"
+    bl_idname = 'color.saffron'
+    def execute(self, context):
+        set_base_color(0xF4C430, self.bl_label)
+        return {'FINISHED'}
+
+class FINALLY_HEALING(bpy.types.Operator):
+    """Finally Healing"""
+    bl_label = "Finally Healing"
+    bl_idname = 'color.finally_healing'
+    def execute(self, context):
+        set_base_color(0xEBE094, self.bl_label)
+        return {'FINISHED'}
+
+class COSMIC_SPECTRUM_GREEN(bpy.types.Operator):
+    """Cosmic Spectrum Green"""
+    bl_label = "Cosmic Spectrum Green"
+    bl_idname = 'color.cosmic_spectrum_green'
+    def execute(self, context):
+        set_base_color(0x9CFFCE, self.bl_label)
+        return {'FINISHED'}
+
+class PHTHALO_GREEN(bpy.types.Operator):
+    """Phthalo Green"""
+    bl_label = "Phthalo Green"
+    bl_idname = 'color.phthalo_green'
+    def execute(self, context):
+        set_base_color(0x123524, self.bl_label)
+        return {'FINISHED'}
+
+class VIRIDIAN(bpy.types.Operator):
+    """Viridian"""
+    bl_label = "Viridian"
+    bl_idname = 'color.viridian'
+    def execute(self, context):
+        set_base_color(0x40826D, self.bl_label)
+        return {'FINISHED'}
+
+class OLO(bpy.types.Operator):
+    """Olo"""
+    bl_label = "Olo"
+    bl_idname = 'color.olo'
+    def execute(self, context):
+        set_base_color(0x00FFCC, self.bl_label)
+        return {'FINISHED'}
+
+class CERULEAN(bpy.types.Operator):
+    """Cerulean"""
+    bl_label = "Cerulean"
+    bl_idname = 'color.cerulean'
+    def execute(self, context):
+        set_base_color(0x00BFFF, self.bl_label)
+        return {'FINISHED'}
+
+class CERULEAN_GREEN(bpy.types.Operator):
+    """Cerulean Green"""
+    bl_label = "Cerulean Green"
+    bl_idname = 'color.cerulean_green'
+    def execute(self, context):
+        set_base_color(0x007BA7, self.bl_label)
+        return {'FINISHED'}
+
+class KLEIN_BLUE(bpy.types.Operator):
+    """Klein Blue"""
+    bl_label = "Klein Blue"
+    bl_idname = 'color.klein_blue'
+    def execute(self, context):
+        set_base_color(0x002FA7, self.bl_label)
+        return {'FINISHED'}
+
+class CERULEAN_BLUE(bpy.types.Operator):
+    """Cerulean Blue"""
+    bl_label = "Cerulean Blue"
+    bl_idname = 'color.cerulean_blue'
+    def execute(self, context):
+        set_base_color(0x2A52BE, self.bl_label)
+        return {'FINISHED'}
+
+class PHTHALO_BLUE(bpy.types.Operator):
+    """Phthalo Blue"""
+    bl_label = "Phthalo Blue"
+    bl_idname = 'color.phthalo_blue'
+    def execute(self, context):
+        set_base_color(0x000F89, self.bl_label)
+        return {'FINISHED'}
+
+class LAVENDER(bpy.types.Operator):
+    """Lavender"""
+    bl_label = "Lavender"
+    bl_idname = 'color.lavender'
+    def execute(self, context):
+        set_base_color(0xE6E6FA, self.bl_label)
+        return {'FINISHED'}
+
+class PERIWINKLE(bpy.types.Operator):
+    """Periwinkle"""
+    bl_label = "Periwinkle"
+    bl_idname = 'color.periwinkle'
+    def execute(self, context):
+        set_base_color(0xCCCCFF, self.bl_label)
+        return {'FINISHED'}
+
+class ULTRMARINE(bpy.types.Operator):
+    """Ultrmarine"""
+    bl_label = "Ultrmarine"
+    bl_idname = 'color.ultrmarine'
+    def execute(self, context):
+        set_base_color(0x120A8F, self.bl_label)
+        return {'FINISHED'}
+
+class INDIGO(bpy.types.Operator):
+    """Indigo"""
+    bl_label = "Indigo"
+    bl_idname = 'color.indigo'
+    def execute(self, context):
+        set_base_color(0x4000FF, self.bl_label)
+        return {'FINISHED'}
+
+class AMETHYST(bpy.types.Operator):
+    """Amethyst"""
+    bl_label = "Amethyst"
+    bl_idname = 'color.amethyst'
+    def execute(self, context):
+        set_base_color(0x9966CC, self.bl_label)
+        return {'FINISHED'}
+
+class LAVENDER_FLORAL(bpy.types.Operator):
+    """Lavender Floral"""
+    bl_label = "Lavender Floral"
+    bl_idname = 'color.lavender_floral'
+    def execute(self, context):
+        set_base_color(0xB57EDC, self.bl_label)
+        return {'FINISHED'}
+
+class WISTERIA(bpy.types.Operator):
+    """Wisteria"""
+    bl_label = "Wisteria"
+    bl_idname = 'color.wisteria'
+    def execute(self, context):
+        set_base_color(0xC9A0DC, self.bl_label)
+        return {'FINISHED'}
+
+class HELIOTROPE(bpy.types.Operator):
+    """Heliotrope"""
+    bl_label = "Heliotrope"
+    bl_idname = 'color.heliotrope'
+    def execute(self, context):
+        set_base_color(0xDF73FF, self.bl_label)
+        return {'FINISHED'}
+
+class DEEPLY_UNCOMFORTABLE(bpy.types.Operator):
+    """Deeply Uncomfortable"""
+    bl_label = "Deeply Uncomfortable"
+    bl_idname = 'color.deeply_uncomfortable'
+    def execute(self, context):
+        set_base_color(0x412A42, self.bl_label)
+        return {'FINISHED'}
+
+class CERISE(bpy.types.Operator):
+    """Cerise"""
+    bl_label = "Cerise"
+    bl_idname = 'color.cerise'
+    def execute(self, context):
+        set_base_color(0xFF00BF, self.bl_label)
+        return {'FINISHED'}
+
+class AUBERGINE(bpy.types.Operator):
+    """Aubergine"""
+    bl_label = "Aubergine"
+    bl_idname = 'color.aubergine'
+    def execute(self, context):
+        set_base_color(0x614051, self.bl_label)
+        return {'FINISHED'}
+
+class BAKER_MILLER_PINK(bpy.types.Operator):
+    """Baker-Miller Pink"""
+    bl_label = "Baker-Miller Pink"
+    bl_idname = 'color.baker_miller_pink'
+    def execute(self, context):
+        set_base_color(0xFF91AF, self.bl_label)
+        return {'FINISHED'}
+
+class CRIMSON(bpy.types.Operator):
+    """Crimson"""
+    bl_label = "Crimson"
+    bl_idname = 'color.crimson'
+    def execute(self, context):
+        set_base_color(0xDC143C, self.bl_label)
+        return {'FINISHED'}
+
+class ALIZARIN_CRIMSON(bpy.types.Operator):
+    """Alizarin Crimson"""
+    bl_label = "Alizarin Crimson"
+    bl_idname = 'color.alizarin_crimson'
+    def execute(self, context):
+        set_base_color(0xE32636, self.bl_label)
+        return {'FINISHED'}
+
+class SAFETY_BLACK(bpy.types.Operator):
+    """Safety Black"""
+    bl_label = "Safety Black"
+    bl_idname = 'color.safety_black'
+    def execute(self, context):
+        set_base_color(0x191919, self.bl_label)
+        return {'FINISHED'}
+
+class REAL_BLACK(bpy.types.Operator):
+    """Real Black"""
+    bl_label = "Real Black"
+    bl_idname = 'color.real_black'
+    def execute(self, context):
+        set_base_color(0x292724, self.bl_label)
+        return {'FINISHED'}
+
+class TRUE_WHITE(bpy.types.Operator):
+    """True White"""
+    bl_label = "True White"
+    bl_idname = 'color.true_white'
+    def execute(self, context):
+        set_base_color(0xE9E9E2, self.bl_label)
+        return {'FINISHED'}
+
+class REAL_WHITE(bpy.types.Operator):
+    """Real White"""
+    bl_label = "Real White"
+    bl_idname = 'color.real_white'
+    def execute(self, context):
+        set_base_color(0xD6D8DB, self.bl_label)
+        return {'FINISHED'}
+
+class TRUE_BLACK(bpy.types.Operator):
+    """True Black (Eigengrau)"""
+    bl_label = "True Black (Eigengrau)"
+    bl_idname = 'color.true_black'
+    def execute(self, context):
+        set_base_color(0x16161D, self.bl_label)
+        return {'FINISHED'}
+
+
+# QMC SELECT PANEL
+
+class QMC SelectPanel(bpy.types.Panel):
+    bl_idname = "QMC SELECT_PT_Panel"
+    bl_label = "QMC Select"
+    bl_space_type = "VIEW_3D"
+    bl_region_type = "UI"
+    bl_category = "MAT"
+    bl_parent_id = 'QMC_PT_Panel'
+    bl_options = {'DEFAULT_CLOSED'}
+
+    def draw(self, context):
+        g.c_icons
+        layout = self.layout
+
+        srow = layout.row()
+        scol = srow.column(align=True)
+        scol.scale_y = 1.25
+        scol.label(text="", icon_value=g.c_icons["international_orange_engineering"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["scarlet"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["vermilion"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["international_orange_aerospace"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["persimmon"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["persian_orange"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["buff"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["cosmic_latte"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["amber"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["saffron"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["finally_healing"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["cosmic_spectrum_green"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["phthalo_green"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["viridian"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["olo"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["cerulean"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["cerulean_green"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["klein_blue"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["cerulean_blue"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["phthalo_blue"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["lavender"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["periwinkle"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["ultrmarine"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["indigo"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["amethyst"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["lavender_floral"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["wisteria"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["heliotrope"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["deeply_uncomfortable"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["cerise"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["aubergine"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["baker_miller_pink"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["crimson"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["alizarin_crimson"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["safety_black"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["real_black"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["true_white"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["real_white"].icon_id)
+        scol.label(text="", icon_value=g.c_icons["true_black"].icon_id)
+
+        scol = srow.column(align=True)
+        scol.scale_y = 1.25
+        scol.scale_x = 3.0
+        scol.operator("color.international_orange_engineering", text="International Orange Engineering")
+        scol.operator("color.scarlet", text="Scarlet")
+        scol.operator("color.vermilion", text="Vermilion")
+        scol.operator("color.international_orange_aerospace", text="International Orange Aerospace")
+        scol.operator("color.persimmon", text="Persimmon")
+        scol.operator("color.persian_orange", text="Persian Orange")
+        scol.operator("color.buff", text="Buff")
+        scol.operator("color.cosmic_latte", text="Cosmic Latte")
+        scol.operator("color.amber", text="Amber")
+        scol.operator("color.saffron", text="Saffron")
+        scol.operator("color.finally_healing", text="Finally Healing")
+        scol.operator("color.cosmic_spectrum_green", text="Cosmic Spectrum Green")
+        scol.operator("color.phthalo_green", text="Phthalo Green")
+        scol.operator("color.viridian", text="Viridian")
+        scol.operator("color.olo", text="Olo")
+        scol.operator("color.cerulean", text="Cerulean")
+        scol.operator("color.cerulean_green", text="Cerulean Green")
+        scol.operator("color.klein_blue", text="Klein Blue")
+        scol.operator("color.cerulean_blue", text="Cerulean Blue")
+        scol.operator("color.phthalo_blue", text="Phthalo Blue")
+        scol.operator("color.lavender", text="Lavender")
+        scol.operator("color.periwinkle", text="Periwinkle")
+        scol.operator("color.ultrmarine", text="Ultrmarine")
+        scol.operator("color.indigo", text="Indigo")
+        scol.operator("color.amethyst", text="Amethyst")
+        scol.operator("color.lavender_floral", text="Lavender Floral")
+        scol.operator("color.wisteria", text="Wisteria")
+        scol.operator("color.heliotrope", text="Heliotrope")
+        scol.operator("color.deeply_uncomfortable", text="Deeply Uncomfortable")
+        scol.operator("color.cerise", text="Cerise")
+        scol.operator("color.aubergine", text="Aubergine")
+        scol.operator("color.baker_miller_pink", text="Baker-Miller Pink")
+        scol.operator("color.crimson", text="Crimson")
+        scol.operator("color.alizarin_crimson", text="Alizarin Crimson")
+        scol.operator("color.safety_black", text="Safety Black")
+        scol.operator("color.real_black", text="Real Black")
+        scol.operator("color.true_white", text="True White")
+        scol.operator("color.real_white", text="Real White")
+        scol.operator("color.true_black", text="True Black (Eigengrau)")
+
+
+# QMC SELECT CLASSES
+array_qmc select = [
+    QMC SelectPanel,
+    INTERNATIONAL_ORANGE_ENGINEERING,
+    SCARLET,
+    VERMILION,
+    INTERNATIONAL_ORANGE_AEROSPACE,
+    PERSIMMON,
+    PERSIAN_ORANGE,
+    BUFF,
+    COSMIC_LATTE,
+    AMBER,
+    SAFFRON,
+    FINALLY_HEALING,
+    COSMIC_SPECTRUM_GREEN,
+    PHTHALO_GREEN,
+    VIRIDIAN,
+    OLO,
+    CERULEAN,
+    CERULEAN_GREEN,
+    KLEIN_BLUE,
+    CERULEAN_BLUE,
+    PHTHALO_BLUE,
+    LAVENDER,
+    PERIWINKLE,
+    ULTRMARINE,
+    INDIGO,
+    AMETHYST,
+    LAVENDER_FLORAL,
+    WISTERIA,
+    HELIOTROPE,
+    DEEPLY_UNCOMFORTABLE,
+    CERISE,
+    AUBERGINE,
+    BAKER_MILLER_PINK,
+    CRIMSON,
+    ALIZARIN_CRIMSON,
+    SAFETY_BLACK,
+    REAL_BLACK,
+    TRUE_WHITE,
+    REAL_WHITE,
+    TRUE_BLACK,
+]